requirements.txt
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
fake-useragent==0.1.11
get==0.0.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
lxml==4.5.1
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.5
post==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygsheets==2.0.3.1
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
query-string==0.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.9
